bdpbinomial Object Plot

plot method for class bdpbinomial.

Usage

# S4 method for class 'bdpbinomial'
plot(x, type = NULL, print = TRUE)

Arguments

x

object of class bdpbinomial. The result of a call to the bdpbinomial function.

type

character. Optional. Select plot type to print. Supports the following: "discount" gives the discount function; "posteriors" gives the posterior plots of the event rates; and "density" gives the augmented posterior density plot(s). Leave NULL to display all plots in sequence.

print

logical. Optional. Produce a plot (print = TRUE; default) or return a ggplot object (print = FALSE). When print = FALSE, it is possible to use ggplot2 syntax to modify the plot appearance.

Details

The plot method for bdpbinomial returns up to three plots: (1) posterior density curves; (2) posterior density of the effect of interest; and (3) the discount function. A call to plot that omits the type input returns one plot at a time and prompts the user to click the plot or press return to proceed to the next plot. Otherwise, the user can specify a plot type to display the requested plot.